Use double instead of float.

+++ src/camera/camera.h
@@ -8,19 +8,19 @@
 class camera {
  public:
   /// t0 から t1で移動するカメラ
-  camera(vec3 lookfrom, vec3 lookat, vec3 vup, float vfov,
-         float aspect, float aperture, float focus_dist,
-         float t0, float t1)
+  camera(vec3 lookfrom, vec3 lookat, vec3 vup, double vfov,
+         double aspect, double aperture, double focus_dist,
+         double t0, double t1)
       : camera(lookfrom, lookat, vup, vfov, aspect, aperture, focus_dist) {
     time0 = t0;
     time1 = t1;
   }
   /// Defocus Blur
-  camera(vec3 lookfrom, vec3 lookat, vec3 vup, float vfov, float aspect, float aperture, float focus_dist) {
+  camera(vec3 lookfrom, vec3 lookat, vec3 vup, double vfov, double aspect, double aperture, double focus_dist) {
     lens_radius = aperture / 2;
-    float theta = vfov * M_PI / 180;
-    float half_height = tan(theta / 2);
-    float half_width = aspect * half_height;
+    double theta = vfov * M_PI / 180;
+    double half_height = tan(theta / 2);
+    double half_width = aspect * half_height;
     origin = lookfrom;
     w = unit_vector(lookfrom - lookat);
     u = unit_vector(cross(vup, w));
@@ -29,10 +29,10 @@ class camera {
     horizontal = 2 * half_width * focus_dist * u;
     vertical = 2 * half_height * focus_dist * v;
   }
-  camera(vec3 lookfrom, vec3 lookat, vec3 vup, float vfov, float aspect) {
-    float theta = vfov * M_PI / 180;
-    float half_height = tan(theta / 2);
-    float half_width = aspect * half_height;
+  camera(vec3 lookfrom, vec3 lookat, vec3 vup, double vfov, double aspect) {
+    double theta = vfov * M_PI / 180;
+    double half_height = tan(theta / 2);
+    double half_width = aspect * half_height;
     origin = lookfrom;
     w = unit_vector(lookfrom - lookat);
     u = unit_vector(cross(vup, w));
@@ -47,11 +47,11 @@ class camera {
     vertical = vec3(0.0, 3.0, 0.0);
     origin = vec3(0.0, 0.0, 0.0);
   }
-  //ray get_ray(float u, float v) {return ray(origin, lower_left_corner + u*horizontal + v*vertical - origin);}
-  ray get_ray(float s, float t) {
+  //ray get_ray(double u, double v) {return ray(origin, lower_left_corner + u*horizontal + v*vertical - origin);}
+  ray get_ray(double s, double t) {
     vec3 rd = lens_radius * random_in_unit_disk();
     vec3 offset = u * rd.x() + v * rd.y();
-    float time = time0 + drand48() * (time1 - time0);
+    double time = time0 + drand48() * (time1 - time0);
     return ray(origin + offset, lower_left_corner + s * horizontal + t * vertical - origin - offset, time);
   }
 
@@ -60,8 +60,8 @@ class camera {
   vec3 horizontal;
   vec3 vertical;
   vec3 u, v, w;
-  float time0{0.0f};
-  float time1{0.0f};
-  float lens_radius;
+  double time0{0.0};
+  double time1{0.0};
+  double lens_radius;
 };
 #endif //RAY_CAMERA_CAMERA_H_
