+++ src/utils/vec3.h
@@ -11,32 +11,32 @@
 class vec3 {
  public:
   vec3() {}
-  vec3(float e0, float e1, float e2) {
+  vec3(double e0, double e1, double e2) {
     e[0] = e0;
     e[1] = e1;
     e[2] = e2;
   }
-  inline float x() const { return e[0]; }
-  inline float y() const { return e[1]; }
-  inline float z() const { return e[2]; }
+  inline double x() const { return e[0]; }
+  inline double y() const { return e[1]; }
+  inline double z() const { return e[2]; }
 
   inline const vec3 &operator+() const { return *this; }
   inline vec3 operator-() const { return vec3(-e[0], -e[1], -e[2]); }
-  inline float operator[](int i) const { return e[i]; }
-  inline float &operator[](int i) { return e[i]; };
+  inline double operator[](int i) const { return e[i]; }
+  inline double &operator[](int i) { return e[i]; };
 
   inline vec3 &operator+=(const vec3 &v2);
   inline vec3 &operator-=(const vec3 &v2);
   inline vec3 &operator*=(const vec3 &v2);
   inline vec3 &operator/=(const vec3 &v2);
-  inline vec3 &operator*=(const float t);
-  inline vec3 &operator/=(const float t);
+  inline vec3 &operator*=(const double t);
+  inline vec3 &operator/=(const double t);
 
-  float length() const {
+  double length() const {
     return sqrt(e[0] * e[0] + e[1] * e[1] + e[2] * e[2]);
   }
 
-  float squared_length() const {
+  double squared_length() const {
     return e[0] * e[0] + e[1] * e[1] + e[2] * e[2];
   }
 
@@ -53,7 +53,7 @@ class vec3 {
     return vec3(random_double(min, max), random_double(min, max), random_double(min, max));
   }
 
-  float e[3];
+  double e[3];
 };
 
 inline std::istream &operator>>(std::istream &is, vec3 &t) {
@@ -82,15 +82,15 @@ inline vec3 operator/(const vec3 &v1, const vec3 &v2) {
   return vec3(v1.e[0] / v2.e[0], v1.e[1] / v2.e[1], v1.e[2] / v2.e[2]);
 }
 
-inline vec3 operator*(float t, const vec3 &v) {
+inline vec3 operator*(double t, const vec3 &v) {
   return vec3(t * v.e[0], t * v.e[1], t * v.e[2]);
 }
 
-inline vec3 operator/(vec3 v, float t) {
+inline vec3 operator/(vec3 v, double t) {
   return vec3(v.e[0] / t, v.e[1] / t, v.e[2] / t);
 }
 
-inline vec3 operator*(const vec3 &v, float t) {
+inline vec3 operator*(const vec3 &v, double t) {
   return vec3(t * v.e[0], t * v.e[1], t * v.e[2]);
 }
 
@@ -122,15 +122,15 @@ inline vec3 &vec3::operator-=(const vec3 &v) {
   return *this;
 }
 
-inline vec3 &vec3::operator*=(const float t) {
+inline vec3 &vec3::operator*=(const double t) {
   e[0] *= t;
   e[1] *= t;
   e[2] *= t;
   return *this;
 }
 
-inline vec3 &vec3::operator/=(const float t) {
-  float k = 1.0 / t;
+inline vec3 &vec3::operator/=(const double t) {
+  double k = 1.0 / t;
 
   e[0] *= k;
   e[1] *= k;
@@ -138,7 +138,7 @@ inline vec3 &vec3::operator/=(const float t) {
   return *this;
 }
 
-inline float dot(const vec3 &v1, const vec3 &v2) {
+inline double dot(const vec3 &v1, const vec3 &v2) {
   return v1.e[0] * v2.e[0] + v1.e[1] * v2.e[1] + v1.e[2] * v2.e[2];
 }
 
@@ -184,6 +184,6 @@ inline vec3 random_in_hemisphere(const vec3 &normal) {
 using point3 = vec3;   // 位置座標
 using color = vec3;    // カラー
 
-static const vec3 ZERO_VEC3(0.0f, 0.0f, 0.0f);
-static const vec3 Y_UP(0.0f, 1.0f, 0.0f);
+static const vec3 ZERO_VEC3(0.0, 0.0, 0.0);
+static const vec3 Y_UP(0.0, 1.0, 0.0);
 #endif //RAY_SRC_UTILS_VEC3_H_
